chore: move languages to separate file
