+++ src/routes/onboarding/user/_languages.ts
@@ -1,0 +1,20 @@
+interface Language {
+	code: string
+	name: string
+}
+
+const languages: Language[] = [
+	{ code: 'af', name: 'Afrikaans (Afrikaans)' },
+	{ code: 'sq', name: 'Albanees (Shqip)' },
+	{ code: 'ar', name: 'Arabisch (العربية)' },
+	{ code: 'hy', name: 'Armeens (Հայերեն)' },
+	{ code: 'as', name: 'Assamees (অসমীয়া)' },
+	{ code: 'de', name: 'Duits (Deutsch)' },
+	{ code: 'en', name: 'Engels (English)' },
+	{ code: 'fr', name: 'Frans (français)' },
+	{ code: 'nl', name: 'Nederlands (Nederlands)' },
+	{ code: 'es', name: 'Spaans (español)' },
+	{ code: 'tr', name: 'Turks (Türkçe)' },
+]
+
+export default languages
